% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjSD2.R
\name{AdjSD2}
\alias{AdjSD2}
\title{Calculates standard deviation of bathymetry (a measure of rugosity) adjusted for slope}
\usage{
AdjSD2(r, w = c(3, 3), na.rm = FALSE, include_scale = FALSE)
}
\arguments{
\item{r}{DEM as a raster layer in a projected coordinate system where map units match elevation/depth units}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}

\item{include_scale}{logical indicating whether to append window size to the layer names (default = FALSE)}
}
\value{
a RasterLayer of adjusted rugosity
}
\description{
Calculates standard deviation of bathymetry (a measure of rugosity). Using a sliding rectangular window a plane is fit to the data and the standard deviation of the residuals is calculated.
}
