% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjusted_Rugosity.R
\name{Adjusted_Rugosity}
\alias{Adjusted_Rugosity}
\title{Calculates standard deviation of bathymetry (a measure of rugosity) adjusted for slope}
\usage{
Adjusted_Rugosity(r, w, na.rm = FALSE, pad = FALSE)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}

\item{pad}{logical value specifying whether rows/columns of NA's should be padded to the edge of the raster to remove edge effects (FALSE by default). If pad is TRUE, na.rm must be TRUE.}
}
\value{
a RasterLayer of adjusted rugosity
}
\description{
Calculates standard deviation of bathymetry (a measure of rugosity). Using a sliding rectangular window a plane is fit to the data and the standard deviation of the residuals is calculated.
}
