% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_code.R
\name{chunk_raster}
\alias{chunk_raster}
\title{Break raster into smaller chunks}
\usage{
chunk_raster(r, n_chunks, buffer)
}
\arguments{
\item{r}{SpatRaster}

\item{n_chunks}{Desired number of chunks to break raster into}

\item{buffer}{the number of rows above/below the cell value that the calculation needs access to. For cell by cell calculations this should be zero and for standard focal operations this would be (w-1)/2 where w is the number of rows in the focal window.}
}
\value{
a list containing a dataframe specifying how rasters are chunked and a list of the chunked SpatRasters
}
\description{
Break raster into smaller chunks
}
