% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_features_ff.R
\name{classify_features_ff}
\alias{classify_features_ff}
\title{IN DEVELOPMENT. Helper function factory to classify morphometric features}
\usage{
classify_features_ff(slope_tolerance, curvature_tolerance)
}
\arguments{
\item{slope_tolerance}{Slope tolerance that defines a 'flat' surface (degrees; default is 1.0). Relevant for the features layer.}

\item{curvature_tolerance}{Curvature tolerance that defines 'planar' surface (default is 0.01). Relevant for the features layer.}
}
\description{
IN DEVELOPMENT. Helper function factory to classify morphometric features according to Wood 1996. Function is not vectorized so it does not work with overlay as of now.
}
