% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_code.R
\name{combine_raster_chunks}
\alias{combine_raster_chunks}
\title{Merge raster chunks from a list into a single raster}
\usage{
combine_raster_chunks(x, breaks_df)
}
\arguments{
\item{x}{A list of SpatRasters}

\item{breaks_df}{Dataframe output from chunk_raster}
}
\value{
a SpatRaster
}
\description{
Merge raster chunks from a list into a single raster
}
