% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aspect.R
\name{convert_aspect}
\alias{convert_aspect}
\title{Helper function to convert aspect to clockwise distance from North}
\usage{
convert_aspect(aspect)
}
\arguments{
\item{aspect}{a number in radians representing aspect as calculated by atan2(dz.dy, -dz.dx)}
}
\value{
aspect as to clockwise distance in radians from North
}
\description{
#' Helper function to convert aspect to clockwise distance from North (0 to 2pi) from counterclockwise distance from east (-pi to pi)
}
\details{
Adapted from https://desktop.arcgis.com/en/arcmap/10.3/tools/spatial-analyst-toolbox/how-aspect-works.htm
}
