% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aspect.R
\name{convert_aspect2}
\alias{convert_aspect2}
\title{Helper function to convert aspect to clockwise distance from North}
\usage{
convert_aspect2(aspect)
}
\arguments{
\item{aspect}{a number in radians representing aspect as calculated by atan2(e,d)}
}
\value{
aspect as to clockwise distance in radians from North
}
\description{
#' Helper function to
}
