% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMV.R
\name{DMV}
\alias{DMV}
\title{Calculates Difference from Mean Value (DMV)}
\usage{
DMV(
  r,
  w = ifelse(tolower(shape) == "rectangle", c(3, 3), NA_real_),
  shape = "rectangle",
  radius = dplyr::case_when(is.matrix(w) ~ NA_real_, tolower(shape) == "circle" &
    tolower(unit) == "cell" ~ 1, tolower(shape) == "circle" & tolower(unit) == "map" ~
    max(terra::res(r)), TRUE ~ NA_real_),
  unit = "cell",
  stand = "none",
  na.rm = FALSE,
  include_scale = FALSE,
  filename = NULL,
  overwrite = FALSE,
  wopt = list()
)
}
\arguments{
\item{r}{DTM as a SpatRaster or RasterLayer}

\item{w}{For a "rectangle" focal window, a vector of length 2 specifying the dimensions where the first number is the number of rows and the second number is the number of columns (or a single number if the number of rows and columns is equal). Window size must be an odd number, and the default is 3x3. w can also be a focal weights matrix. Circular windows this can be created using MultiscaleDTM::circle_window or w can be set to NULL and radius can be used instead.}

\item{shape}{character representing the shape of the focal window. Either "rectangle" (default), or "circle".}

\item{radius}{A single integer representing the radius of the circle or a vector in "cell" or "map" units. This is ignored if w is not NULL.}

\item{unit}{unit for radius. Either "cell" (number of cells, the default) or "map" for map units (e.g. meters).}

\item{stand}{standardization method. Either "none" (the default), "range" or "sd" indicating whether the relative position should be standardized by the standard deviation or range of values in the focal window. If stand is 'none' the layer name will be dmv, otherwise it will be sdmv to indicate that the layer has been standardized.}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}

\item{include_scale}{logical indicating whether to append window size to the layer names (default = FALSE). If unit="map" then window size will have "MU" after the number indicating that the number represents the window size in map units if a circular window is used.}

\item{filename}{character Output filename.}

\item{overwrite}{logical. If TRUE, filename is overwritten (default is FALSE).}

\item{wopt}{list with named options for writing files as in writeRaster}
}
\value{
a SpatRaster or RasterLayer
}
\description{
Calculates Difference from Mean Value (DMV) which is a measure of relative position. DMV calculates the difference between the focal cell and the mean of all cells in a rectangular or circular neighborhood. Positive values indicate local highs (i.e. peaks) and negative values indicate local lows (i.e. depressions). DMV can be expressed in units of the input DTM raster or can standardized relative to the local topography by dividing by the standard deviation or range of elevation values in the focal window.
}
\examples{
r<- rast(volcano, extent= ext(2667400, 2667400 + 
ncol(volcano)*10, 6478700, 6478700 + nrow(volcano)*10), 
crs = "EPSG:27200")
dmv<- DMV(r, w=c(5,5), shape= "rectangle", na.rm = TRUE, stand="range")
plot(dmv)
}
\references{
Lecours, V., Devillers, R., Simms, A.E., Lucieer, V.L., Brown, C.J., 2017. Towards a Framework for Terrain Attribute Selection in Environmental Studies. Environmental Modelling & Software 89, 19-30. https://doi.org/10.1016/j.envsoft.2016.11.027
Wilson, J.P., Gallant, J.C. (Eds.), 2000. Terrain Analysis: Principles and Applications. John Wiley & Sons, Inc.
}
