% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erupt.R
\name{erupt}
\alias{erupt}
\title{Create georeferenced version of R's built in volcano dataset}
\usage{
erupt()
}
\value{
SpatRaster
}
\description{
Create georeferenced version of R's built in volcano dataset. Useful dataset for generating quick examples.
}
\examples{
r<- erupt()
}
