% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_code.R
\name{focalCpp_parallel}
\alias{focalCpp_parallel}
\title{FocalCpp for use in parallel code}
\usage{
focalCpp_parallel(x, ...)
}
\arguments{
\item{x}{a packed SpatRaster}

\item{...}{Other arguments for terra::focalCpp}
}
\value{
a packed SpatRaster
}
\description{
FocalCpp for use in parallel code
}
