% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvatures.R
\name{ku}
\alias{ku}
\title{Calculate unsphericity curvature}
\usage{
ku(a, b, c, d, e)
}
\arguments{
\item{a}{regression coefficient}

\item{b}{regression coefficient}

\item{c}{regression coefficient}

\item{d}{regression coefficient}

\item{e}{regression coefficient}
}
\description{
Calculate  unsphericity curvature, ku, from the equation Z =ax^2+by^2+cxy+dx+ey(+f).
}
\references{
Evans, I.S., 1980. An integrated system of terrain analysis and slope mapping. Zeitschrift f¨ur Geomorphologic Suppl-Bd 36, 274–295.

Wood, J., 1996. The geomorphological characterisation of digital elevation models (Ph.D.). University of Leicester.

Minár, J., Evans, I.S., Jenčo, M., 2020. A comprehensive system of definitions of land surface (topographic) curvatures, with implications for their application in geoscience modelling and prediction. Earth-Science Reviews 211, 103414. https://doi.org/10.1016/j.earscirev.2020.103414
}
