% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qfit.R
\name{outlier_filter}
\alias{outlier_filter}
\title{Helper function to filter outliers from regression parameters using interquartile range}
\usage{
outlier_filter(params, outlier_quantile)
}
\arguments{
\item{params}{regression parameters for fitted surface}

\item{outlier_quantile}{vector of length 2 specifying the quantiles used for filtering outliers}
}
\description{
Helper function to filter outliers from regression parameters using interquartile range
}
