% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeAspect.R
\name{SlopeAspect}
\alias{SlopeAspect}
\title{Multiscale Slope and Aspect}
\usage{
SlopeAspect(
  r,
  w = c(3, 3),
  unit = "degrees",
  method = "queen",
  metrics = c("slope", "aspect", "northness", "eastness"),
  include_scale = FALSE,
  mask_aspect = TRUE
)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{unit}{"degrees" or "radians"}

\item{method}{"queen" or "rook", indicating how many neighboring cells to use to compute slope for any cell. queen uses 8 neighbors (up, down, left, right, and diagonals) and rook uses 4 (up, down, left, right).}

\item{metrics}{a character string or vector of character strings of which terrain atrributes to return ("slope" and/or "aspect"). Default is c("slope", "aspect").}

\item{include_scale}{logical indicating whether to append window size to the layer names (default = FALSE)}

\item{mask_aspect}{A logical. If slope evaluates to 0, aspect will be set to NA when mask_aspect is TRUE (the default). If FALSE, when slope is 0 aspect will be pi/2 radians or 90 degrees which is the behavior of raster::terrain.}
}
\value{
a RasterStack or RasterLayer of slope and/or aspect
}
\description{
Calculates multiscale slope and aspect based on the slope.k/aspect.k algorithm from Misiuk et al (2021) which extends classical formulations of slope restricted to a 3x3 window to multiple scales. The code from Misiuk et al (2021) was modified to allow for rectangular rather than only square windows.
}
\details{
When method="rook", slope and aspect are computed according to Fleming and Hoffer (1979) and Ritter (1987). When method="queen", slope and aspect are computed according to Horn (1981). These are the standard slope algorithms found in many GIS packages but are traditionally restricted to a 3 x 3 window size. Misiuk et al (2021) extended these classical formulations  to multiple window sizes. This function modifies the code from Misiuk et al (2021) to allow for rectangular rather than only square windows and also added aspect.
}
