% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfaceArea2.R
\name{SurfaceArea2}
\alias{SurfaceArea2}
\title{Calculates surface area of a DEM}
\usage{
SurfaceArea2(r, run_in_blocks)
}
\arguments{
\item{r}{DEM as a raster layer}
}
\value{
a RasterLayer
}
\description{
Calculates surface area on a per cell basis of a DEM based on Jenness, 2004. This wrapper for sp::surfaceArea that natively works on rasters.
}
