% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRM.R
\name{VRM}
\alias{VRM}
\title{Implementation of the Sappington et al., (2007) vector ruggedness measure}
\usage{
VRM(r, w, include_scale = FALSE)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{include_scale}{logical indicating whether to append window size to the layer names (default = FALSE)}
}
\value{
a RasterLayer
}
\description{
Implementation of the Sappington et al., (2007) vector ruggedness measure, modified from Evans (2021).
}
\references{
Evans JS (2021). spatialEco. R package version 1.3-6, https://github.com/jeffreyevans/spatialEco.

Sappington, J.M., Longshore, K.M., Thompson, D.B., 2007. Quantifying Landscape Ruggedness for Animal Habitat Analysis: A Case Study Using Bighorn Sheep in the Mojave Desert. The Journal of Wildlife Management 71, 1419-1426. https://doi.org/10.2193/2005-723
}
