% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodEvans.R
\name{WoodEvans}
\alias{WoodEvans}
\title{Calculates multiscale slope and aspect}
\usage{
WoodEvans(
  r,
  w,
  unit = "degrees",
  return_aspect = FALSE,
  slope_tolerance = 1,
  curvature_tolerance = 1e-04,
  na.rm = FALSE,
  pad = FALSE,
  return_params = FALSE
)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{unit}{"degrees" or "radians"}

\item{return_aspect}{logical indicating whether or not to return aspect in addition to Northness and Eastness (default is FALSE)}

\item{slope_tolerance}{Slope tolerance that defines a 'flat' surface (degrees; default is 1.0). Relevant for the features layer.}

\item{curvature_tolerance}{Curvature tolerance that defines 'planar' surface (default is 0.0001). Relevant for the features layer.}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}

\item{pad}{logical value specifying whether rows/columns of NA's should be padded to the edge of the raster to remove edge effects (FALSE by default). If pad is TRUE, na.rm must be TRUE.}

\item{return_params}{logical indicating whether to return Wood/Evans regression parameters (default FALSE)}
}
\value{
a RasterStack
}
\description{
Calculates Calculates multiscale slope and aspect of a DEM over a sliding rectangular window using a quadratic fit to the surface according to the Wood/Evans procedure. This is an R/C++ implementation of r.param.scale GRASS GIS function. For aspect 0 degrees represents north and increases clockwise which differs from the way r.param.scale reports aspect.
}
