% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodEvansHelper.R
\name{WoodEvansHelper}
\alias{WoodEvansHelper}
\title{Helper function to calculates Wood/Evans regression parameters}
\usage{
WoodEvansHelper(r, w, type, na.rm)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{type}{1 (linear) or 2 (quadratic). Default is type 2}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}
}
\value{
a RasterStack of slope, eastness, northness, and optionally aspect
}
\description{
Helper function to calculates Wood/Evans regression parameters
}
