% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodEvansHelper2.R
\name{WoodEvansHelper2}
\alias{WoodEvansHelper2}
\title{Helper function to calculates Wood/Evans regression parameters}
\usage{
WoodEvansHelper2(r, w, type, resolution, na.rm)
}
\arguments{
\item{r}{DEM as a matrix}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{type}{1 (linear) or 2 (quadratic). Default is type 2}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}
}
\value{
a list containing matrices of regression parameters (a-f) and the standard deviation of residuals
}
\description{
Helper function to calculates Wood/Evans regression parameters
}
