% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPI.R
\name{BPI}
\alias{BPI}
\title{Calculates Bathymetric Position Index}
\usage{
BPI(
  r,
  radius,
  stand = "none",
  unit = "cell",
  w = NULL,
  na.rm = FALSE,
  include_scale = FALSE,
  filename = NULL,
  overwrite = FALSE,
  wopt = list()
)
}
\arguments{
\item{r}{r DTM as a SpatRaster or RasterLayer.}

\item{radius}{Vector of length 2 specifying c(inner, outer) radii of the annulus c(inner,outer) in "cell" or "map" units. Inner radius must be less than or equal to outer radius.}

\item{stand}{Standardization method. Either "none" (the default), "range" or "sd" indicating whether the relative position should be standardized by dividing by the standard deviation or range of included values in the focal window. If stand is 'none' the layer name will be "bpi", otherwise it will be "sbpi" to indicate that the layer has been standardized.}

\item{unit}{Unit for radius. Either "cell" (number of cells, the default) or "map" for map units (e.g. meters).}

\item{w}{A focal weights matrix created MultiscaleDTM::annulus_window or w can be set to NA or NULL and radius can be used instead.}

\item{na.rm}{A logical vector indicating whether or not to remove NA values before calculations}

\item{include_scale}{Logical indicating whether to append window size to the layer names (default = FALSE). If include_scale = TRUE the inner and outer radius will be appended to the layer name. If unit="map" then window size will have "MU" after the number indicating that the number represents the scale in map units.}

\item{filename}{Character output filename.}

\item{overwrite}{Logical. If TRUE, filename is overwritten (default is FALSE).}

\item{wopt}{List with named options for writing files as in writeRaster.}
}
\value{
A SpatRaster or RasterLayer.
}
\description{
Calculates Bathymetric Position Index (BPI). BPI is a measure of relative position that calculates the difference between the value of the focal cell and the mean of cells contained within an annulus shaped neighborhood. Positive values indicate local highs (i.e. peaks) and negative values indicate local lows (i.e. depressions). BPI can be expressed in units of the input DTM raster or can standardized relative to the local topography by dividing by the standard deviation or range of included elevation values in the focal window.
}
\examples{
r<- rast(volcano, extent= ext(2667400, 2667400 + 
ncol(volcano)*10, 6478700, 6478700 + nrow(volcano)*10), 
crs = "EPSG:27200")
bpi<- BPI(r, radius = c(2,4), stand= "none", unit = "cell", na.rm = TRUE)
plot(bpi)
}
\references{
Lundblad, E.R., Wright, D.J., Miller, J., Larkin, E.M., Rinehart, R., Naar, D.F., Donahue, B.T., Anderson, S.M., Battista, T., 2006. A benthic terrain classification scheme for American Samoa. Marine Geodesy 29, 89–111. https://doi.org/10.1080/01490410600738021
}
