% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfit.R
\name{Pfit}
\alias{Pfit}
\title{Calculates multiscale slope and aspect using a local planar fit (IN DEVELOPMENT: DO NOT USE)}
\usage{
Pfit(
  r,
  w = c(3, 3),
  unit = "degrees",
  na.rm = FALSE,
  include_scale = FALSE,
  filename = NULL,
  overwrite = FALSE,
  wopt = list()
)
}
\arguments{
\item{r}{DTM as a SpatRaster (terra) or RasterLayer (raster) in a projected coordinate system where map units match elevation/depth units (up is assumed to be north for calculations of aspect, northness, and eastness).}

\item{w}{Vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number. Default is 3x3.}

\item{unit}{"degrees" or "radians".}

\item{na.rm}{Logical indicating whether or not to remove NA values before calculations.}

\item{filename}{character Output filename. Can be a single filename, or as many filenames as there are layers to write a file for each layer}

\item{overwrite}{logical. If TRUE, filename is overwritten (default is FALSE).}

\item{wopt}{list with named options for writing files as in writeRaster}
}
\value{
a SpatRaster (terra) or RasterStack/RasterLayer (raster)
}
\description{
Calculates multiscale slope and aspect using a local planar fit (IN DEVELOPMENT: DO NOT USE)
}
\references{
Sharpnack, D. A. and Akin, G., 1969. An algorithm for computing slope and aspect from elevations. Photogrammetric Engineering 35(3), 247-248.
}
