% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfaceArea.R
\name{SurfaceArea}
\alias{SurfaceArea}
\title{Calculates surface area of a DEM}
\usage{
SurfaceArea(r, expand = FALSE)
}
\arguments{
\item{r}{DEM as a SpatRaster or RasterLayer in a projected coordinate system where map units match elevation/depth units}
}
\value{
a SpatRaster or RasterLayer
}
\description{
Calculates surface area on a per cell basis of a DEM based on Jenness, 2004.
}
\references{
Jenness, J.S., 2004. Calculating landscape surface area from digital elevation models. Wildlife Society Bulletin 32, 829-839. https://doi.org/10.2193/0091-7648(2004)032\link{0829:CLSAFD}2.0.CO;2
}
