% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRM.R
\name{VRM}
\alias{VRM}
\title{Implementation of the Sappington et al., (2007) vector ruggedness measure}
\usage{
VRM(r, w, include_scale)
}
\arguments{
\item{r}{DEM as a raster layer}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{include_scale}{logical indicating whether to append window size to the layer names (default = FALSE)}
}
\value{
a RasterLayer
}
\description{
Implementation of the Sappington et al., (2007) vector ruggedness measure. This is a simple wrapper around the vrm function from spatialEco by Jeffrey S Evans, to make the syntax consistent with the rest of the MultiscaleDEM package, and is included in this package since VRM is also a rugosity measure that corrects for slope. If you use this function, please cite the spatialEco package.
}
